% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion-tools.R
\name{adaptive_gkernel}
\alias{adaptive_gkernel}
\title{Adaptive Gaussian kernel of matrix}
\usage{
adaptive_gkernel(x, k = 2, cutoff = Inf, alpha = 2)
}
\arguments{
\item{x}{Matrix of distances}

\item{k}{K-th neighbor distances (used for kernal width)}

\item{cutoff}{Maximum kernel width}

\item{alpha}{Kernel power}
}
\value{
Matrix
}
\description{
Adaptive Gaussian kernel of matrix
}
