% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion-tools.R
\name{affinity}
\alias{affinity}
\title{Calculate affinity matrix}
\usage{
affinity(m, k = 2, cutoff = Inf, alpha = 2)
}
\arguments{
\item{m}{Matrix of distances}

\item{k}{K-th neighbor distances (used for kernal width)}

\item{cutoff}{Maximum kernel width}

\item{alpha}{Kernel power}
}
\value{
Matrix
}
\description{
Uses adaptive Gaussian kernel.
}
