% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-tidying.R
\name{as_df}
\alias{as_df}
\alias{as_df.data.frame}
\alias{as_df.matrix}
\alias{as_df.logical}
\alias{as_df.character}
\alias{as_df.numeric}
\title{Cast to tibble}
\usage{
as_df(x, ...)

\method{as_df}{data.frame}(x, rownames = "rownames", ...)

\method{as_df}{matrix}(x, rownames = "rownames", ...)

\method{as_df}{logical}(x, ...)

\method{as_df}{character}(x, ...)

\method{as_df}{numeric}(x, ...)
}
\arguments{
\item{x}{Data frame}

\item{x}{Matrix}
}
\value{
Data frame
}
\description{
Cast to tibble
}
