% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{colorscheme}
\alias{colorscheme}
\title{Eachan's color scheme}
\usage{
colorscheme(color = NULL)
}
\arguments{
\item{color}{Color name}
}
\value{
Hex color
}
\description{
Return either  a list of colors, or match first argument to single color and return that as hex.
}
