% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-tidying.R
\name{.as_df_v}
\alias{.as_df_v}
\title{Turn a vector into a data frame}
\usage{
.as_df_v(x, key_name = "key", value_name = "value", keys = NULL, ...)
}
\arguments{
\item{x}{Vector}

\item{key_name}{Column name for keys}

\item{value_name}{Column name for values}

\item{keys}{If x is unnamed, use these as keys}
}
\description{
Turn a vector into a data frame.
}
