% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion-tools.R
\name{get_k_nn_dist}
\alias{get_k_nn_dist}
\title{Get distance of k nearest neighbor}
\usage{
get_k_nn_dist(x, k, cutoff = Inf)
}
\arguments{
\item{x}{Matrix of distances}

\item{k}{K-th neighbor}

\item{cutoff}{Maximum distance to return}
}
\value{
1-length numeric vector
}
\description{
Get distance of k nearest neighbor
}
