% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion-tools.R
\name{gkernel}
\alias{gkernel}
\title{Gaussian kernel of matrix}
\usage{
gkernel(x, sigma = 1, alpha = 2)
}
\arguments{
\item{x}{Matrix of distances}

\item{sigma}{Kernel width}

\item{alpha}{Kernel power}
}
\value{
Matrix
}
\description{
Gaussian kernel of matrix
}
