% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%?\%}
\alias{\%?\%}
\title{Ternary functions}
\usage{
x \%?\% y
}
\arguments{
\item{x}{Condition which evaluates to \code{logical}}

\item{y}{Truth value}
}
\value{
Object of class \code{ternary}
}
\description{
Ternary functions.
}
