% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%or\%}
\alias{\%or\%}
\title{Failback operator}
\usage{
x \%or\% y
}
\arguments{
\item{x}{Value which might be 0-ish}

\item{y}{Failback value}
}
\value{
LHS or RHS
}
\description{
Binary operator: if the left-hand-side is \code{NULL}, \code{0}, \code{FALSE}, or \code{NA},
returns the right-hand-side. Otherwise returns the left-hand-side.
}
