% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{heat_map}
\alias{heat_map}
\alias{heat_map.matrix}
\alias{heat_map.data.frame}
\title{Plot a heat map}
\usage{
heat_map(x, ...)

\method{heat_map}{matrix}(x, filename = NULL, width = 12,
  height = 12, hclustfun = function(x) hclust(x, "ward.D2"),
  shrink_labels = c(1, 1), heat_palette = "RdBu", row_palette = NULL,
  col_palette = NULL, ...)

\method{heat_map}{data.frame}(x, filename = NULL, width = 12,
  height = 12, hclustfun = function(x) hclust(x, "ward.D2"),
  shrink_labels = c(1, 1), heat_palette = "RdBu", row_palette = NULL,
  col_palette = NULL, ...)
}
\arguments{
\item{x}{A matrix, Matricks, or data frame}

\item{...}{Passed to \code{gplots::heatmap.2}}

\item{filename}{File to save as}

\item{width}{File width}

\item{height}{File height}

\item{hclustfun}{Function for clustering}

\item{shrink_labels}{Length-2 numeric vector for how much to shrink row and column labels}

\item{heat_palette}{RColorBrewer palette for values}

\item{row_palette}{RColorBrewer palette for row side-information}

\item{col_palette}{RColorBrewer palette for colum side-information}

\item{x}{A data frame}
}
\value{
A plot

A plot
}
\description{
Plot a heat map
}
