% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{log_plot}
\alias{log_plot}
\alias{log_plot.gg}
\title{Add log axis}
\usage{
log_plot(x, ...)

\method{log_plot}{gg}(x, axes = "x")
}
\arguments{
\item{x}{A data frame}

\item{axes}{Which axes to log-transform}
}
\value{
\code{ggplot2} object
}
\description{
Add log axis
}
