% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion-tools.R
\name{make_transition_matrix}
\alias{make_transition_matrix}
\title{Calculate Markov matrix}
\usage{
make_transition_matrix(m, k = 2, npcs = 1, cutoff = Inf, alpha = 2,
  distfun = function(x) parallelDist::parDist(x, threads =
  system("nproc") - 1))
}
\arguments{
\item{m}{Matrix}

\item{k}{K-th neighbor distances (used for kernal width)}

\item{cutoff}{Maximum kernel width}

\item{alpha}{Kernel power}

\item{distfun}{Function to calculate distnace between rows}
}
\value{
Row-normalized Markov matrix
}
\description{
Calculate Markov matrix
}
