% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_cols}
\alias{map_cols}
\title{Column data manipulation}
\usage{
map_cols(x, columns, new_column_name = "new_column", fun = function(x)
  x)
}
\arguments{
\item{x}{Data frame}
}
\value{
Data frame
}
\description{
Column data manipulation
}
\examples{
d <- data.frame(A=LETTERS, B=1:26, C=letters)
map_cols(d, c('A', 'B'), fun=function(x) apply(x, 1, paste, collapse='__'))

}
