% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matricks.R
\name{Matricks}
\alias{Matricks}
\title{Matricks}
\usage{
Matricks(data = NA, rows_title = NULL, cols_title = NULL,
  rows_annotation = NULL, cols_annotation = NULL, ...)
}
\arguments{
\item{data}{Matrix}

\item{rows_title}{Title or ID for rownames}

\item{cols_title}{Title or ID for colnames}

\item{rows_annotation}{Data frame of anntotations for rows. Must have a column name in common with rows_title}

\item{cols_annotation}{Data frame of anntotations for columns. Must have a column name in common with cols_title.}
}
\value{
Matricks object; inherits from \code{matrix}.
}
\description{
Matrix with side information. See \link{matrix} for description of parameters.
}
\examples{
annotation_df <- data.frame(col_id=LETTERS[1:3], annotation=letters[1:3])
m <- matrix(rnorm(9), ncol=3)
colnames(m) <- annotation_df$col_id
Matricks(m, cols_annotation=annotation_df)
}
