% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packEJ-package.R
\docType{package}
\name{packEJ}
\alias{packEJ}
\alias{packEJ-package}
\title{packEJ: Eachan Johnson's personal R package.}
\description{
Helpful functions I use over and over.
}
\section{Plotting functions}{

Plotting functions include dose repsonses, heatmaps, and dimension reduction visualization.
}

