% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-tidying.R
\name{pivot_long}
\alias{pivot_long}
\alias{pivot_long.matrix}
\alias{pivot_long.data.frame}
\alias{pivot_long.matricks}
\title{Pivot data structure to long}
\usage{
pivot_long(x, ...)

\method{pivot_long}{matrix}(x, varnames = names(dimnames(x)), ...,
  na.rm = FALSE, as.is = FALSE, value.name = "value")

\method{pivot_long}{data.frame}(x, id.vars = NULL, measure.vars = NULL,
  variable.name = "key", ..., na.rm = FALSE, value.name = "value",
  factorsAsStrings = TRUE)

\method{pivot_long}{matricks}(x, varnames = names(dimnames(x)), ...,
  na.rm = FALSE, as.is = FALSE, value.name = "value")
}
\arguments{
\item{x}{Data structure}
}
\value{
Data frame in long format
}
\description{
Force tidy format
}
\examples{
pivot_long(matrix(rnorm(9), 3), rownames='observation')
pivot_long(data.frame(A=rnorm(9), B=LETTERS[1:9], C=letters[1:9]), rownames='observation', id.vars='B')
annotation_df <- data.frame(col_id=LETTERS[1:3], annotation=letters[1:3])
m <- matrix(rnorm(9), ncol=3)
colnames(m) <- annotation_df$col_id
m2 <- Matricks(m, cols_annotation=annotation_df)
pivot_long(m2)
}
