% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-tidying.R
\name{pivot_wide}
\alias{pivot_wide}
\alias{pivot_wide.data.frame}
\title{Pivot data structure to wide}
\usage{
pivot_wide(x, ...)

\method{pivot_wide}{data.frame}(x, key.vars, id.vars, value.var,
  rows_annotation = NULL, cols_annotation = NULL,
  fun.aggregate = mean, ..., fill = NULL, drop = TRUE)
}
\arguments{
\item{x}{Data frame}
}
\value{
Data frame in wide format Matricks
}
\description{
Force wide (matrix) format
}
\examples{
annotation_df <- data.frame(col_id=LETTERS[1:3], annotation=letters[1:3])
m <- matrix(rnorm(9), ncol=3)
colnames(m) <- annotation_df$col_id
m2 <- Matricks(m, cols_annotation=annotation_df)
d <- pivot_long(m2)
pivot_wide(d, 'col_id', 'rownames', 'value', cols_annotation='annotation')

}
