% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dose_response}
\alias{plot_dose_response}
\alias{plot_dose_response.data.frame}
\title{Plot a dose response curve}
\usage{
plot_dose_response(x, ...)

\method{plot_dose_response}{data.frame}(x, xvar, yvar, wt = "H37Rv",
  highlights = NULL, group = "paste(strain, compound)",
  facet = "compound_name", bg_color = "grey80", wt_color = "grey40",
  palette = "Dark2", highlight_size = 1, alpha = 1,
  labeller = ggplot2::as_labeller(function(x) x), scales = "free_x",
  filename = NULL, panel_width = 2, panel_height = 2,
  legend_width = 1)
}
\arguments{
\item{x}{A data frame}

\item{xvar}{X-axis column}

\item{yvar}{Y-axis column}

\item{wt}{String to match for WT strains}

\item{highlights}{Data frame for filtering high;ighted conditions}

\item{group}{Grouping}

\item{facet}{Faceting}

\item{bg_color}{Color for bulk of strains}

\item{wt_color}{Color for WT strains}

\item{palette}{Palette for highlighted lines}

\item{highlight_size}{Size of highlighted interaction lines}

\item{filename}{Filename to save}

\item{panel_width}{Width of each panel in inches}

\item{panel_height}{Height of each panel in inches}

\item{legend_width}{Width of legend in inches}
}
\value{
ggplot2 object
}
\description{
Plot a dose response curve
}
