% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matricks.R
\name{print.matricks}
\alias{print.matricks}
\title{Show a Matricks}
\usage{
\method{print}{matricks}(x, ...)
}
\arguments{
\item{x}{\code{\link{Matricks}} object.}
}
\description{
Pretty printing
}
\examples{
annotation_df <- data.frame(col_id=LETTERS[1:3], annotation=letters[1:3])
m <- matrix(rnorm(9), ncol=3)
colnames(m) <- annotation_df$col_id
m2 <- Matricks(m, cols_annotation=annotation_df)
print(m2)

}
