% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{println}
\alias{println}
\title{Print with timestamp}
\usage{
println(...)
}
\arguments{
\item{...}{Values to concatenate and print to console using \code{cat}.}
}
\description{
Wrapper for \code{cat} which adds a timestamp and newline.
}
