% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion-tools.R
\name{row_normalize}
\alias{row_normalize}
\title{Normalize matrix rows}
\usage{
row_normalize(m, rowfun = rowSums)
}
\arguments{
\item{m}{Matrix}

\item{rowfun}{Function to calculate normalization denominator}
}
\value{
Matrix
}
\description{
Normalize matrix rows
}
