% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{set_theme}
\alias{set_theme}
\title{Set Eachan's default plotting theme}
\usage{
set_theme(font_size = 15, font = "Coolvetica Rg")
}
\arguments{
\item{font_size}{Font size to use}

\item{font}{Font to use}
}
\description{
Set Eachan's default plotting theme
}
