% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matricks.R
\name{[.matricks}
\alias{[.matricks}
\title{Subset a matricks}
\usage{
\method{[}{matricks}(x, i, j, drop = missing(i) \%or\% length(cols) == 1)
}
\arguments{
\item{x}{\code{\link{Matricks}} object.}

\item{i}{Row indices}

\item{j}{Column indicies}

\item{drop}{Not implemented}
}
\description{
Subset matrix and annotations on the fly
}
\examples{
annotation_df <- data.frame(col_id=LETTERS[1:3], annotation=letters[1:3])
m <- matrix(rnorm(9), ncol=3)
colnames(m) <- annotation_df$col_id
m2 <- Matricks(m, cols_annotation=annotation_df)
print(m2[1:3, ])

}
