% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matricks.R
\name{subset.matricks}
\alias{subset.matricks}
\alias{[.matricks}
\title{Subset a matricks}
\usage{
\method{subset}{matricks}(x, i, j, drop = missing(i) \%or\% length(cols)
  == 1, filter = NULL)

\method{[}{matricks}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{\code{\link{Matricks}} object.}

\item{i}{Row indices}

\item{j}{Column indicies}

\item{drop}{As usual for [}

\item{filter}{A data frame for a filtering join on side-information, or a named vector of
column names and values to keep.}

\item{i}{Row indices}

\item{j}{Column indicies}

\item{drop}{As usual for [}
}
\description{
Subset matrix and annotations on the fly
}
\examples{
annotation_df <- data.frame(col_id=LETTERS[1:3], annotation=letters[1:3])
m <- matrix(rnorm(9), ncol=3)
colnames(m) <- annotation_df$col_id
m2 <- Matricks(m, cols_annotation=annotation_df)
filter_df <- data.frame(col_id=LETTERS[1:2])
print(subset(m2, filter=filter_df))

print(subset(m2, filter=list(col_id=LETTERS[1:2])))

annotation_df <- data.frame(col_id=LETTERS[1:3], annotation=letters[1:3])
m <- matrix(rnorm(9), ncol=3)
colnames(m) <- annotation_df$col_id
m2 <- Matricks(m, cols_annotation=annotation_df)
print(m2[1:3, ])

}
