% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{heat_map}
\alias{heat_map}
\alias{heat_map.matrix}
\alias{heat_map.data.frame}
\title{Plot a heat map}
\usage{
heat_map(x, ...)

\method{heat_map}{matrix}(x, filename = NULL, width = 12,
  height = 12, hclustfun = function(x) hclust(x, "ward.D2"),
  shrink_labels = c(1, 1), heat_palette = "RdBu", row_palette = NULL,
  col_palette = NULL, ...)

\method{heat_map}{data.frame}(x, filename = NULL, width = 12,
  height = 12, hclustfun = function(x) hclust(x, "ward.D2"),
  shrink_labels = c(1, 1), heat_palette = "RdBu", row_palette = NULL,
  col_palette = NULL, ...)
}
\arguments{
\item{x}{A matrix}

\item{x}{A data frame}
}
\value{
A plot

A plot
}
\description{
Plot a heat map

Plot a heat map.
}
