% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join}
\alias{join}
\alias{join.character}
\title{Concatenate a character vector}
\usage{
join(x, ...)

\method{join}{character}(x, char = ", ", ...)
}
\arguments{
\item{x}{Object to join}

\item{x}{Character vector}
}
\description{
Concatenate a character vector
}
