% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dose_response}
\alias{plot_dose_response}
\alias{plot_dose_response.data.frame}
\title{Plot a dose response curve}
\usage{
plot_dose_response(x, ...)

\method{plot_dose_response}{data.frame}(x, xvar, yvar, wt = "H37Rv",
  highlights = NULL, group = "paste(strain, compound)",
  facet = "compound_name", bg_color = "grey80", wt_color = "grey40",
  palette = "Dark2", highlight_size = 1, alpha = 1,
  labeller = ggplot2::as_labeller(function(x) x), scales = "free_x",
  filename = NULL, panel_width = 2, panel_height = 2,
  legend_width = 1)
}
\arguments{
\item{x}{A data frame}

\item{xvar}{X-axis column}

\item{yvar}{Y-axis column}

\item{legend_width}{}
}
\value{
ggplot2 object
}
\description{
Create a new factor from two existing factors, where the new factor's levels
are the union of the levels of the input factors.
}
